Add CSS to stop photos inverting in Windows high contrast mode
In Windows High Contrast mode, the photos invert, and it makes them look a bit scary! This makes sure that all the MOJ Blocks images don't invert when this mode is activated on a Windows machine.

I haven't been able to test yet, as it isn't something you can test in browser stack (it requires being able to modify settings on an operating system level). I do have a Windows laptop though, so can test when this goes onto dev.

I've put this change into the Blocks, but I'm aware it might affect photos that people upload in ways that don't use the blocks. So I'm open to changing the approach!