Adjust block margins left and right
Created an override to remove the left right margins added in by the Gov
UK framework. This is so all our blocks line up within the Hale them
main container. Once we refactor our theme we can revisit this override
in our block depending on the framework we choose for the theme.