Refactor hero block to be dynamically loaded
As part of our overall stratagy to make sure what we develop will be
compatible with future releases this turns the default static block
design to a dynamic one. This has involved adding a PHP generated
frontend, rather then the WP react frontend that is saved into the DB
and then output.